test: move Codex auth fingerprint apply coverage
